Document most of tor-proto
(except for the worst parts that need refactoring the most)

+++ tor-proto/src/crypto/cell.rs
@@ -1,34 +1,74 @@
+//! Relay cell cryptography
+//!
+//! The Tor protocol centers around "RELAY cells", which are
+//! transmitted through the network along circuits.  The client that
+//! creates a circuitg shares two different set of keys and state with
+//! each of the relays on the circuit: one for "oubound" traffic, and
+//! one for "inbound" traffic.
+//!
+
 use crate::{Error, Result};
 
+/// The number of bytes in the body of a cell.
+///
+/// TODO: This probably belongs somewhere else.
 pub const CELL_BODY_LEN: usize = 509;
+/// A cell body considerd as a raw array of bytes
 pub type RawCellBody = [u8; CELL_BODY_LEN];
+/// Type for a cell body.
 #[derive(Clone)]
 pub struct CellBody(pub RawCellBody);
 
+/// Represents the ability for a circuit crypto state to be initialized
+/// from a given seed.
 pub trait CryptInit {
+    /// Return the number of bytes that this state will require.
     fn seed_len() -> usize;
+    /// Construct this state from a seed of the appropriate length.
+    ///
+    /// TODO: maybe this should return a Result rather than just
+    /// asserting that the length is correct.
     fn initialize(seed: &[u8]) -> Self;
 }
 
+/// Represents a relay's view of the crypto state on a given circuit.
 pub trait RelayCrypt {
+    /// Prepare a CellBody to be sent towards the client.
     fn originate(&mut self, cell: &mut CellBody);
+    /// Encrypt a CellBody that is moving towards the client.
     fn encrypt_inbound(&mut self, cell: &mut CellBody);
+    /// Decrypt a CellBody that is moving towards the client.
+    ///
+    /// Return true if it is addressed to us.
     fn decrypt_outbound(&mut self, cell: &mut CellBody) -> bool;
 }
 
+/// A client's view of the crypto state shared with a single relay.
 pub trait ClientLayer {
+    /// Prepare a CellBody to be sent to the relay at this layer.
     fn originate_for(&mut self, cell: &mut CellBody);
+    /// Encrypt a CellBody to be decrypted by this layer.
     fn encrypt_outbound(&mut self, cell: &mut CellBody);
+    /// Decrypt a CellBopdy that passed through this layer.
+    ///
+    /// Return true if this layer is the originator.
     fn decrypt_inbound(&mut self, cell: &mut CellBody) -> bool;
 }
 
-pub type HopNum = u16;
+/// Type to store hop indices on a circuit.
+pub type HopNum = u8;
 
+/// A client's view of the cryptographic state for an entire
+/// constructed circuit.
 pub struct ClientCrypt {
     layers: Vec<Box<dyn ClientLayer>>,
 }
 
 impl ClientCrypt {
+    /// Prepare a cell body to sent away from the client.
+    ///
+    /// The cell is prepared for the `hop`th hop, and then encrypted with
+    /// the appropriate keys.
     pub fn encrypt(&mut self, cell: &mut CellBody, hop: HopNum) -> Result<()> {
         let hop = hop as usize;
         if hop > self.layers.len() {
@@ -41,6 +81,9 @@ impl ClientCrypt {
         }
         Ok(())
     }
+    /// Decrypt an incoming cell that is coming to the client.
+    ///
+    /// On success, return which hop was the originator of the cell.
     pub fn decrypt(&mut self, cell: &mut CellBody) -> Result<HopNum> {
         for (hopnum, layer) in self.layers.iter_mut().enumerate() {
             if layer.decrypt_inbound(cell) {
@@ -49,12 +92,14 @@ impl ClientCrypt {
         }
         Err(Error::BadCellAuth)
     }
+    /// Add a new layer to this ClientCrypt
     pub fn add_layer(&mut self, layer: Box<dyn ClientLayer>) {
         assert!(self.layers.len() < HopNum::max_value() as usize);
         self.layers.push(layer);
     }
 }
 
+/// Incomplete untested implementation of Tor's current cell crypto.
 mod tor1 {
     use super::*;
     use digest::Digest;
@@ -62,6 +107,11 @@ mod tor1 {
     use stream_cipher::{NewStreamCipher, StreamCipher};
     use typenum::Unsigned;
 
+    /// A CryptState is part of a RelayCrypt or a ClientLayer.
+    ///
+    /// It is parameterized on a stream cipher and a digest type: most
+    /// circuits will use AES-128-CTR and SHA1, but v3 onion services
+    /// use AES-256-CTR and SHA-3.
     pub struct CryptState<SC: StreamCipher, D: Digest + Clone> {
         f_c: SC,
         b_c: SC,
@@ -117,6 +167,7 @@ mod tor1 {
     }
 
     impl CellBody {
+        /// Prepare a cell body by setting its digest and recognized field.
         fn set_digest<D: Digest + Clone>(&mut self, d: &mut D) {
             self.0[1] = 0;
             self.0[2] = 0;
@@ -129,6 +180,7 @@ mod tor1 {
             let r = d.clone().result(); // XXX can I avoid this clone?
             self.0[5..9].copy_from_slice(&r[0..4]);
         }
+        /// Check a cell to see whether its recognized field is set.
         fn recognized<D: Digest + Clone>(&mut self, d: &mut D) -> bool {
             // maybe too optimized? XXXX
             // XXXX self is only mut for an optimization.
